+++ lib/mamba.dart
@@ -1,6 +1,7 @@
 /// APIs for defining, parsing, and executing Mamba commands.
 library;
 
+export 'command.dart';
 export 'context.dart';
 export 'errors.dart';
 export 'executor.dart';